tests/wpt/web-platform-tests/shadow-dom/untriaged/elements-and-dom-objects/shadowroot-object/shadowroot-methods/test-010.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_10_01_02_09</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="https://dom.spec.whatwg.org/#dom-node-clonenode">
<meta name="assert" content="If context object is a shadow root, throw a NotSupportedError exception.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../../../html/resources/common.js"></script>
<script src="../../../../resources/shadow-dom-utils.js"></script>
</head>
<body>
<div id="log"></div>
<script>
test(unit(function (ctx) {
    var d = newRenderedHTMLDocument(ctx);
    var host = d.createElement('div');
    d.body.appendChild(host);
    var s = host.attachShadow({mode: 'open'});

    try {
        s.cloneNode();
        assert_true(false, 'Invoking the cloneNode() method on a ShadowRoot instance must always ' +
        'throw a NotSupportedError (code 9) exception.');
    } catch (e) {
        assert_equals(e.code, 9, 'Wrong exception type');
        assert_equals(e.name, 'NotSupportedError', 'Wrong exception name');
    }
}), 'A_10_01_02_09_T01');
</script>
</body>
</html>
